--- app/selectors/errorSelector.js
@@ -1,13 +1,0 @@
-/**
- * If an error happened, it's in here
- */
-
-import { createSelector } from 'reselect';
-import globalSelector from './globalSelector';
-
-const errorSelector = () => createSelector(
-  globalSelector(),
-  (globalState) => globalState.get('error')
-);
-
-export default errorSelector;
