--- app/selectors/globalSelector.js
@@ -1,7 +1,0 @@
-/**
- * Select the entire state
- */
-
-const globalSelector = () => (state) => state.get('global');
-
-export default globalSelector;
