--- app/selectors/homeSelector.js
@@ -1,7 +1,0 @@
-/**
- * Select the entire state
- */
-
-const homeSelector = () => (state) => state.get('home');
-
-export default homeSelector;
