--- app/selectors/routeSelector.js
@@ -1,3 +1,0 @@
-const routeSelector = () => (state) => state.get('route');
-
-export default routeSelector;
