--- app/selectors/selectLocationSelector.js
@@ -1,18 +1,0 @@
-// selectLocationState expects a plain JS object for the routing state
-const createSelectLocationState = () => {
-  let prevRoutingState;
-  let prevRoutingStateJS;
-
-  return (state) => {
-    const routingState = state.get('route'); // or state.route
-
-    if (!routingState.equals(prevRoutingState)) {
-      prevRoutingState = routingState;
-      prevRoutingStateJS = routingState.toJS();
-    }
-
-    return prevRoutingStateJS;
-  };
-};
-
-export default createSelectLocationState;
